// src/api/gamification/enums/gamification.enum.ts
export enum RewardTransactionType {
  VERIFICATION_COMPLETED = 'VERIFICATION_COMPLETED',
  LESSON_COMPLETED = 'LESSON_COMPLETED',
  QUIZ_PASSED = 'QUIZ_PASSED',
  CHALLENGE_COMPLETED = 'CHALLENGE_COMPLETED',
  MISSION_COMPLETED = 'MISSION_COMPLETED',
  DAILY_STREAK = 'DAILY_STREAK',
  BADGE_EARNED = 'BADGE_EARNED',
  ADMIN_ADJUSTMENT = 'ADMIN_ADJUSTMENT',
  REVERSAL = 'REVERSAL',
  EVIDENCE_INSPECTED = 'EVIDENCE_INSPECTED',
}

export enum BadgeCriteriaType {
  VERIFICATION_COUNT = 'VERIFICATION_COUNT',
  LESSON_COUNT = 'LESSON_COUNT',
  QUIZ_SCORE = 'QUIZ_SCORE',
  CHALLENGE_STREAK = 'CHALLENGE_STREAK',
  DAILY_STREAK = 'DAILY_STREAK',
  REPORT_FEEDBACK = 'REPORT_FEEDBACK',
  SPECIAL_EVENT = 'SPECIAL_EVENT',
  EVIDENCE_INSPECTION_COUNT = 'EVIDENCE_INSPECTION_COUNT',
  TAGGED_ACTIVITY_COUNT = 'TAGGED_ACTIVITY_COUNT',
  GUIDED_INVESTIGATION_COUNT = 'GUIDED_INVESTIGATION_COUNT',
  MISSION_COUNT = 'MISSION_COUNT',
}

export enum RankCode {
  NOVICE = 'NOVICE',
  EXPLORER = 'EXPLORER',
  INVESTIGATOR = 'INVESTIGATOR',
  VERIFIER = 'VERIFIER',
  TRUTH_CHAMPION = 'TRUTH_CHAMPION',
}

export enum AchievementEventType {
  BADGE_EARNED = 'BADGE_EARNED',
  RANK_UP = 'RANK_UP',
}

export enum BadgeAvailability {
  AVAILABLE = 'AVAILABLE',
  COMING_SOON = 'COMING_SOON',
}

export enum LeaderboardPeriod {
  WEEKLY = 'WEEKLY',
  MONTHLY = 'MONTHLY',
  ALL_TIME = 'ALL_TIME',
}